# tooling/testdata/fixture/apps/api/internal/gen/contoh/contoh.gen.go
// Package contoh provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package contoh

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Contoh defines model for Contoh.
type Contoh struct {
	DibuatPada time.Time          `json:"dibuatPada"`
	Id         openapi_types.UUID `json:"id"`
	Nama       string             `json:"nama"`
}

// ContohCreateRequest Badan permintaan — field opsional memakai *optional*, TIDAK PERNAH nullable (C-04).
type ContohCreateRequest struct {
	Nama string `json:"nama"`
}

// ContohMetaData defines model for ContohMetaData.
type ContohMetaData struct {
	// Pagination Komponen paginasi BERSAMA — satu bentuk untuk seluruh operasi list.
	Pagination Pagination `json:"pagination"`
}

// GetContohParams defines parameters for GetContoh.
type GetContohParams struct {
	Page  *int `form:"page,omitempty" json:"page,omitempty"`
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostContohJSONRequestBody defines body for PostContoh for application/json ContentType.
type PostContohJSONRequestBody = ContohCreateRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetContoh Daftar contoh, berpaginasi
	// (GET /contoh)
	GetContoh(c *gin.Context, params GetContohParams)
	// PostContoh Buat contoh baru
	// (POST /contoh)
	PostContoh(c *gin.Context)
	// GetContohById Ambil satu contoh lewat ID
	// (GET /contoh/{contohId})
	GetContohById(c *gin.Context, contohId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetContoh operation middleware
func (siw *ServerInterfaceWrapper) GetContoh(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetContohParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", c.Request.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", c.Request.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetContoh(c, params)
}

// PostContoh operation middleware
func (siw *ServerInterfaceWrapper) PostContoh(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostContoh(c)
}

// GetContohById operation middleware
func (siw *ServerInterfaceWrapper) GetContohById(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "contohId" -------------
	var contohId string

	err = runtime.BindStyledParameterWithOptions("simple", "contohId", c.Param("contohId"), &contohId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter contohId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetContohById(c, contohId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/contoh", wrapper.GetContoh)
	router.POST(options.BaseURL+"/contoh", wrapper.PostContoh)
	router.GET(options.BaseURL+"/contoh/:contohId", wrapper.GetContohById)
}

type GetContohRequestObject struct {
	Params GetContohParams
}

type GetContohResponseObject interface {
	VisitGetContohResponse(w http.ResponseWriter) error
}

type GetContoh200JSONResponse struct {
	Data []Contoh `json:"data"`

	// Message Kalimat untuk manusia, bukan untuk cabang logika klien.
	Message  string         `json:"message"`
	MetaData ContohMetaData `json:"metaData"`

	// Status Status envelope sukses — nilai tunggal, dan itu disengaja (lihat C-01).
	Status EnvelopeStatus `json:"status"`
}

func (response GetContoh200JSONResponse) VisitGetContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetContoh401JSONResponse EnvelopeError

func (response GetContoh401JSONResponse) VisitGetContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetContoh403JSONResponse EnvelopeError

func (response GetContoh403JSONResponse) VisitGetContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type PostContohRequestObject struct {
	Body *PostContohJSONRequestBody
}

type PostContohResponseObject interface {
	VisitPostContohResponse(w http.ResponseWriter) error
}

type PostContoh201JSONResponse struct {
	Data Contoh `json:"data"`

	// Message Kalimat untuk manusia, bukan untuk cabang logika klien.
	Message string `json:"message"`

	// MetaData metaData membawa HANYA paginasi (C-01) — field kedua apa pun dilarang.
	MetaData *EnvelopeMeta `json:"metaData,omitempty"`

	// Status Status envelope sukses — nilai tunggal, dan itu disengaja (lihat C-01).
	Status EnvelopeStatus `json:"status"`
}

func (response PostContoh201JSONResponse) VisitPostContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PostContoh400JSONResponse EnvelopeError

func (response PostContoh400JSONResponse) VisitPostContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PostContoh401JSONResponse EnvelopeError

func (response PostContoh401JSONResponse) VisitPostContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type PostContoh403JSONResponse EnvelopeError

func (response PostContoh403JSONResponse) VisitPostContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type PostContoh409JSONResponse EnvelopeError

func (response PostContoh409JSONResponse) VisitPostContohResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type GetContohByIdRequestObject struct {
	ContohId string `json:"contohId"`
}

type GetContohByIdResponseObject interface {
	VisitGetContohByIdResponse(w http.ResponseWriter) error
}

type GetContohById200JSONResponse struct {
	Data Contoh `json:"data"`

	// Message Kalimat untuk manusia, bukan untuk cabang logika klien.
	Message string `json:"message"`

	// MetaData metaData membawa HANYA paginasi (C-01) — field kedua apa pun dilarang.
	MetaData *EnvelopeMeta `json:"metaData,omitempty"`

	// Status Status envelope sukses — nilai tunggal, dan itu disengaja (lihat C-01).
	Status EnvelopeStatus `json:"status"`
}

func (response GetContohById200JSONResponse) VisitGetContohByIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetContohById401JSONResponse EnvelopeError

func (response GetContohById401JSONResponse) VisitGetContohByIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetContohById403JSONResponse EnvelopeError

func (response GetContohById403JSONResponse) VisitGetContohByIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type GetContohById404JSONResponse EnvelopeError

func (response GetContohById404JSONResponse) VisitGetContohByIdResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetContoh Daftar contoh, berpaginasi
	// (GET /contoh)
	GetContoh(ctx context.Context, request GetContohRequestObject) (GetContohResponseObject, error)
	// PostContoh Buat contoh baru
	// (POST /contoh)
	PostContoh(ctx context.Context, request PostContohRequestObject) (PostContohResponseObject, error)
	// GetContohById Ambil satu contoh lewat ID
	// (GET /contoh/{contohId})
	GetContohById(ctx context.Context, request GetContohByIdRequestObject) (GetContohByIdResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetContoh operation middleware
func (sh *strictHandler) GetContoh(ctx *gin.Context, params GetContohParams) {
	var request GetContohRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetContoh(ctx, request.(GetContohRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetContoh")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetContohResponseObject); ok {
		if err := validResponse.VisitGetContohResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostContoh operation middleware
func (sh *strictHandler) PostContoh(ctx *gin.Context) {
	var request PostContohRequestObject

	var body PostContohJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(ctx, err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostContoh(ctx, request.(PostContohRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostContoh")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(PostContohResponseObject); ok {
		if err := validResponse.VisitPostContohResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetContohById operation middleware
func (sh *strictHandler) GetContohById(ctx *gin.Context, contohId string) {
	var request GetContohByIdRequestObject

	request.ContohId = contohId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetContohById(ctx, request.(GetContohByIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetContohById")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetContohByIdResponseObject); ok {
		if err := validResponse.VisitGetContohByIdResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}